requirements.txt
alpaca-trade-api
requests
yfinance
python-dotenv
google-genai